Rename events package controller -> ui
