+++ src/main/java/seedu/address/commons/events/ui/ExitAppRequestEvent.java
@@ -1,15 +1,14 @@
-package seedu.address.commons.events.controller;
+package seedu.address.commons.events.ui;
 
 import seedu.address.commons.events.BaseEvent;
 
 /**
- * An event requesting to view the help page.
+ *
  */
-public class ShowHelpEvent extends BaseEvent {
+public class ExitAppRequestEvent extends BaseEvent {
 
     @Override
     public String toString() {
         return this.getClass().getSimpleName();
     }
-
 }
