+++ src/main/java/seedu/address/commons/events/ui/JumpToListRequestEvent.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.events.controller;
+package seedu.address.commons.events.ui;
 
 import seedu.address.commons.events.BaseEvent;
