+++ src/main/java/seedu/address/logic/commands/ExitCommand.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
 import seedu.address.commons.core.EventsCenter;
-import seedu.address.commons.events.controller.ExitAppRequestEvent;
+import seedu.address.commons.events.ui.ExitAppRequestEvent;
 
 /**
  * Terminates the program.
