+++ src/main/java/seedu/address/ui/StatusBarFooter.java
@@ -19,7 +19,7 @@ import seedu.address.commons.events.model.AddressBookChangedEvent;
 import seedu.address.commons.util.FxViewUtil;
 
 /**
- * A controller for the status bar that is displayed at the footer of the application.
+ * A ui for the status bar that is displayed at the footer of the application.
  */
 public class StatusBarFooter extends UiPart {
     private static final Logger logger = LogsCenter.getLogger(StatusBarFooter.class);
