+++ src/main/java/seedu/address/ui/UiManager.java
@@ -10,9 +10,9 @@ import seedu.address.MainApp;
 import seedu.address.commons.core.ComponentManager;
 import seedu.address.commons.core.Config;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.commons.events.controller.JumpToListRequestEvent;
-import seedu.address.commons.events.controller.PersonPanelSelectionChangedEvent;
-import seedu.address.commons.events.controller.ShowHelpEvent;
+import seedu.address.commons.events.ui.JumpToListRequestEvent;
+import seedu.address.commons.events.ui.PersonPanelSelectionChangedEvent;
+import seedu.address.commons.events.ui.ShowHelpEvent;
 import seedu.address.commons.events.storage.DataReadingExceptionEvent;
 import seedu.address.commons.events.storage.DataSavingExceptionEvent;
 import seedu.address.commons.util.StringUtil;
