+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -8,8 +8,8 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.logic.commands.*;
-import seedu.address.commons.events.controller.JumpToListRequestEvent;
-import seedu.address.commons.events.controller.ShowHelpEvent;
+import seedu.address.commons.events.ui.JumpToListRequestEvent;
+import seedu.address.commons.events.ui.ShowHelpEvent;
 import seedu.address.commons.events.model.AddressBookChangedEvent;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
